+++ src/auth/mod.rs
@@ -8,6 +8,8 @@ pub mod user;
 pub use authorizer::{request_action, request_grant};
 #[cfg(feature = "axum")]
 pub use authorizer::authorizer;
+#[cfg(feature = "axum")]
+pub use tower_http::csrf::{ConfigError, CsrfLayer, ProtectionError};
 pub use grant::Grant;
 pub use oauth::{
     build_authorization_url, AuthUrlOptions, BasicClient, BasicTokenResponse, BasicTokenType,